feat: 完成商品能力扩展与V4.4架构加固

+++ src/main/java/com/wms/system/entity/InventoryBatch.java
@@ -264,7 +264,7 @@ public class InventoryBatch extends BaseEntity {
     @Version
     @Column(name = "version", nullable = false)
     @Builder.Default
-    private Long version = 0L;
+    private Integer version = 0;
 
     // ========== 便利方法 (Helper Methods) ==========
